Add a custom User-Agent for usage measurement
The default `User-Agent` for `Net::HTTP` is `Ruby`, so usage
metrics for this gem blend in with all other Ruby code that
is interacting with the Linode API and not setting a custom
value for the `User-Agent` header.

Setting a custom `User-Agent` that includes the gem name and
version will make it easier to get accurate usage metrics as
we move forward.